+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Deprecated
 ### Removed
 ### Fixed
+- Fixed a bug where the LiveLossPlot callback would trigger an error if run and evaluate were called separately
 
 ## [0.2.5] - 2018-12-19
 ### Added